+++ inst/scripts/ResultsBayes_rps.R
@@ -42,7 +42,11 @@ tree$labels <-
   )
 
 # Assign multi- and single-taxa node labels.
-tree$haplotypes <- Thesis::haplotype_labels(haplotypes = tree$joined)
+tree$haplotypes <-
+  Thesis::haplotype_labels(
+    haplotypes = tree$joined,
+    id_column = "Taxon_a_posteriori"
+  )
 tree$single <- tree$joined %>%
   dplyr::group_by(label) %>%
   dplyr::mutate(n = dplyr::n()) %>%
@@ -168,7 +172,6 @@ tree$pdf <- tree$repel_pdf +
       override.aes = list(size = 3),
       title.position = "top",
       ncol = 3,
-      byrow = TRUE,
       keyheight = 0.15,
       default.unit = "inch"
     )
@@ -257,7 +260,10 @@ tree$png <- tree$repel_png +
     color = ggplot2::guide_legend(
       override.aes = list(size = 3),
       title.position = "top",
-      ncol = 1, byrow = TRUE, keyheight = 0.15, default.unit = "inch")
+      ncol = 1,
+      keyheight = 0.15,
+      default.unit = "inch"
+    )
   ) +
   ggplot2::theme(
     legend.position = "right",